+++ src/main/java/chessEngine/chess/piece/constantMovesPiece/king/BlackKing.java
@@ -5,7 +5,7 @@ import chessEngine.chess.piece.PieceColor;
 public class BlackKing extends King {
     private final PieceColor pieceColor = PieceColor.BLACK;
 
-    public BlackKing(int height, int width) {
+    public BlackKing(byte height, byte width) {
         super(height, width);
     }
 }
